GeomGraphSpace: a ggplot2 prototype for GraphSpace-class methods

GeomGraphSpace is the underlying ggproto object used by geom_graphspace to draw node and edge elements in a graph layout.

This geom is designed for network diagrams, where graph attributes are often already in their final form (e.g., hex colors).

Usage

GeomGraphSpace

Aesthetics

GeomGraphSpace understands ggplot2's conventions for point-like geoms.

See also